refactor(webchat): code quality cleanup and component extraction (#628)
- Remove unused imports: newSessionId, urlWorker, httpBase
- Remove unused params: env in handleDelta/handleToolCall/handleToolResult
- Remove dead code: detailsStr/eventStr in runtime adapter
- Remove unused useRef import in admin page
- Delete stale package-lock.json (project uses pnpm)
- Migrate deprecated ThreadPrimitive.Empty to useAuiState isEmpty check
- Export TODO_TOOLS from tool-categories.ts, deduplicate in runtime adapter
- Extract thread.tsx (597 lines) into 7 focused component files:
  AssistantMessage, UserMessage, WelcomeScreen, PreAssistantIndicator,
  ReasoningBlock, CopyButton, MessageActions + shared thread-helpers
