+++ webchat/lib/ai-sdk-transport/client/browser-client.ts
@@ -50,7 +50,6 @@ import {
   createWorkerCommandEnvelope,
   serializeEnvelope,
   deserializeEnvelope,
-  newSessionId,
   isInitAck,
 } from './envelope';
